backend/requirements.txt
fastapi==0.115.0
mangum==0.19.0
uvicorn[standard]==0.32.0
supabase==2.10.0
google-generativeai==0.8.5
python-dotenv==1.0.1
requests==2.32.3
